+++ src/test/resources/fixtures/sorting/svelte/script-helper-template-literal/input.svelte
@@ -1,0 +1,5 @@
+<script>
+  const button = cn(`text-center p-4 flex bg-red-500 font-bold`)
+</script>
+
+<button class={button}>Save</button>