+++ src/test/resources/fixtures/sorting/svelte/sveltekit-component-props/expected.svelte
@@ -1,0 +1,17 @@
+<script>
+  let {
+    class: className = "",
+    compact = false,
+    active = false,
+    ...rest
+  } = $props()
+
+  const classes = cn([
+    "flex bg-red-500 p-4 text-center font-bold",
+    active && "flex p-4 font-bold",
+    { "px-2 py-0.5 text-xs tracking-wide": compact },
+    className,
+  ])
+</script>
+
+<button {...rest} class={classes}>Save</button>